Rename fake protobufs file
